# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/resources": {
      "get": {
        "operationId": "ResourcesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number, starting at 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": ["ASC", "DESC"]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field. Each controller restricts the allowed values, because\ninterpolating a client-supplied column name into ORDER BY is an injection\nvector that no parameter binding protects against.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typeId",
            "required": false,
            "in": "query",
            "description": "Only resources of this type.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Searches name and code.",
            "schema": {
              "example": "aurora",
              "type": "string"
            }
          },
          {
            "name": "minCapacity",
            "required": false,
            "in": "query",
            "description": "Minimum capacity.",
            "schema": {
              "minimum": 1,
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Planta 2",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "By default both active and inactive are shown.",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ResourceDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List resources with filters and pagination",
        "tags": ["Resources"]
      },
      "post": {
        "description": "Registers a bookable resource under an existing type.",
        "operationId": "ResourcesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDto"
                }
              }
            }
          },
          "409": {
            "description": "A resource with that code already exists"
          },
          "422": {
            "description": "The time zone is not valid"
          }
        },
        "summary": "Create a resource",
        "tags": ["Resources"]
      }
    },
    "/api/v1/resources/{id}": {
      "get": {
        "operationId": "ResourcesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDto"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist"
          }
        },
        "summary": "Get one resource, with its type and weekly schedule",
        "tags": ["Resources"]
      },
      "patch": {
        "description": "code and resourceTypeId are not editable: both identify the resource.",
        "operationId": "ResourcesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDto"
                }
              }
            }
          }
        },
        "summary": "Update a resource",
        "tags": ["Resources"]
      },
      "delete": {
        "description": "A soft delete. The resource can no longer be booked, but its reservation history is kept intact -- which is the reason it is not deleted.",
        "operationId": "ResourcesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Deactivate a resource",
        "tags": ["Resources"]
      }
    },
    "/api/v1/resources/{id}/activation": {
      "post": {
        "operationId": "ResourcesController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDto"
                }
              }
            }
          }
        },
        "summary": "Reactivate a deactivated resource",
        "tags": ["Resources"]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Includes how many confirmed future reservations each one has, so the impact of banning or demoting is visible before doing it.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number, starting at 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": ["ASC", "DESC"]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field. Each controller restricts the allowed values, because\ninterpolating a client-supplied column name into ORDER BY is an injection\nvector that no parameter binding protects against.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Searches name and email.",
            "schema": {
              "example": "ana",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List users",
        "tags": ["Users"]
      },
      "post": {
        "description": "Created by an administrator, bypassing public sign-up. Delegated to the identity provider, which is what knows how to hash the password. A provider that cannot create accounts answers 501.",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "An account with that email already exists"
          }
        },
        "summary": "Create an account",
        "tags": ["Users"]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "description": "The email is the credential this person signs in with, so changing it changes how they get in. The role is changed through its own route, because it has its own rule: nobody can demote themselves.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "The email is already used by another account"
          }
        },
        "summary": "Edit name and email",
        "tags": ["Users"]
      }
    },
    "/api/v1/users/{id}/role": {
      "patch": {
        "description": "An administrator cannot remove their own role: it would be the quickest way to leave the system with nobody able to administer it.",
        "operationId": "UsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "Attempted self-demotion"
          }
        },
        "summary": "Change a user's role",
        "tags": ["Users"]
      }
    },
    "/api/v1/users/{id}/ban": {
      "post": {
        "description": "Takes effect immediately: the adapter rejects a banned user on every request, not only at sign-in, and their open sessions are revoked. Without `days` the ban is indefinite.",
        "operationId": "UsersController_ban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "Attempted self-ban"
          }
        },
        "summary": "Ban a user",
        "tags": ["Users"]
      },
      "delete": {
        "operationId": "UsersController_unban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Unban a user",
        "tags": ["Users"]
      }
    },
...
  },
  "info": {
    "title": "Shared Resource Booking API",
    "description": "Booking API for shared resources.\n\n**Central rule:** two `CONFIRMED` reservations never overlap on the same resource.\nIntervals are half-open `[start, end)`, so 10:00–11:00 and 11:00–12:00 are\ncontiguous and both valid.\n\nErrors follow RFC 7807 (`application/problem+json`).",
    "version": "0.3.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by POST /api/auth/sign-in/email"
      }
    },
    "schemas": {
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137
          },
          "totalPages": {
            "type": "number",
            "example": 7
          }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Declared explicitly as an untyped array.\n\nSwagger cannot resolve `T` at runtime, and letting the CLI plugin infer\nit makes the schema self-referential -- the app fails to boot with\n\"circular dependency detected (property key: data)\". The real item type\nis composed by ApiPaginatedResponse() below, via allOf.",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": ["data", "meta"]
      },
      "ResourceTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "meeting-room"
          },
          "name": {
            "type": "string",
            "example": "Sala de reuniones"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": ["id", "code", "name", "isActive"]
      },
      "ResourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "aurora"
          },
          "name": {
            "type": "string",
            "example": "Sala Aurora"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "example": 12
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "Planta 2"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA zone. The client formats times in this zone, not the browser's.",
            "example": "Europe/Madrid"
          },
          "isActive": {
            "type": "boolean"
          },
          "resourceType": {
            "$ref": "#/components/schemas/ResourceTypeDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "code", "name", "timeZone", "isActive", "createdAt"]
      },
      "CreateResourceDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "description": "Maximum capacity. Checked against each reservation's attendees.",
            "minimum": 1,
            "example": 12
          },
          "location": {
            "type": "string",
            "maxLength": 160,
            "example": "Planta 2"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA zone, for example Europe/Madrid.\n\nOperating hours are interpreted in this zone; instants are always stored\nin UTC.",
            "maxLength": 64,
            "example": "Europe/Madrid",
            "default": "UTC"
          },
          "resourceTypeId": {
            "type": "string",
            "description": "The type it belongs to.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique, stable code. Lowercase letters, digits and hyphens.",
            "maxLength": 60,
            "pattern": "^[a-z0-9-]+$"
          },
          "name": {
            "type": "string",
            "description": "Display name.",
            "maxLength": 160
          }
        },
        "required": ["resourceTypeId", "code", "name"]
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "description": "Maximum capacity. Checked against each reservation's attendees.",
            "minimum": 1,
            "example": 12
          },
          "location": {
            "type": "string",
            "maxLength": 160,
            "example": "Planta 2"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA zone, for example Europe/Madrid.\n\nOperating hours are interpreted in this zone; instants are always stored\nin UTC.",
            "maxLength": 64,
            "example": "Europe/Madrid",
            "default": "UTC"
          },
          "name": {
            "type": "string",
            "description": "Display name.",
            "maxLength": 160
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": ["admin", "user"],
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "banned": {
            "type": "boolean"
          },
          "activeReservations": {
            "type": "number",
            "description": "Future confirmed reservations. Shown so an admin sees the impact of a ban.",
            "example": 3
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "emailVerified",
          "banned",
          "activeReservations",
          "createdAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "At least 10 characters, the same as public sign-up.",
            "minLength": 10
          },
          "role": {
            "description": "Defaults to `user`.",
            "enum": ["admin", "user"],
            "type": "string",
            "default": "user"
          },
          "email": {
            "type": "string",
            "description": "Email, unique across the system.",
            "format": "email"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": ["password", "email", "name"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name.",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "description": "Email. It must remain unique.\n\nChanging it changes the credential this person signs in with, which is\nwhy the interface warns before saving.",
            "format": "email"
          }
        }
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "description": "The new role. Only `admin` or `user`.",
            "enum": ["admin", "user"],
            "type": "string"
          }
        },
        "required": ["role"]
      },
      "BanUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for the ban. It is recorded.",
            "maxLength": 300,
            "example": "Misuse of the meeting rooms"
          },
          "days": {
            "type": "number",
            "description": "Duration in days. With no value the ban is indefinite.\n\nBetter Auth lifts the ban itself once it expires, so no scheduled job is\nneeded to check for it.",
            "minimum": 1,
            "example": 30
          }
        }
      }
    }
  }
}